backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
opencv-python==4.8.1.78
ultralytics==8.0.222
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0